backend/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
groq==0.4.2
gunicorn==21.2.0
python-dotenv==1.0.0
